test(toolkit): refactor tests to run with zoneless change detection
